src/todo/urls.py
from django.urls import include, path
from rest_framework.routers import DefaultRouter

from .views import TodoGroupViewSet, TodoViewSet

router = DefaultRouter()
router.register(r"groups", TodoGroupViewSet, basename="todogroup")
router.register(r"todos", TodoViewSet, basename="todo")

urlpatterns = [
    path("", include(router.urls)),
]
