requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9.0
python-decouple>=3.8
djangorestframework>=3.14.0
